% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aids}
\alias{aids}
\title{The effects of AZT in slowing the development of AIDS symptoms}
\format{
A data frame with 4 rows and 4 variables:
\itemize{
\item \code{symptomatic}: counts of veterans showing AIDS symptoms during the 3-year study
\item \code{asymptomatic}: counts of veterans not showing AIDS symptoms during the 3-year study
\item \code{race}: the race of the veterans with levels \code{"White"} and \code{"Black"}
\item \code{AZT_use}: whether the veterans received AZT immediately (\code{"Yes"})
or waited until their T cells showed severe immune weakness (\code{"No"})
}
}
\source{
The data set is analyzed in Agresti (2002, Subsection 5.4.2). Its
original source is New York Times, Feb. 15, 1991.
}
\usage{
aids
}
\description{
The data is from a 3-year study on the effects of AZT in slowing the
development of AIDS symptoms. 338 veterans whose immune systems
were beginning to falter after infection with the AIDS virus were
randomly assigned either to receive AZT immediately or to wait
until their T cells showed severe immune weakness.
}
\references{
Agresti A (2002). \emph{Categorical Data Analysis}.  Wiley Series in
Probability and Statistics. Wiley.
}
\seealso{
\code{\link[=brmultinom]{brmultinom()}}
}
\keyword{datasets}
