% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracl.R
\name{bracl}
\alias{bracl}
\title{Bias reduction for adjcacent category logit models using the
Poisson trick.}
\usage{
bracl(formula, data, weights, subset, na.action, parallel = FALSE,
  contrasts = NULL, model = TRUE, x = TRUE, control = list(...),
  ...)
}
\arguments{
\item{formula}{
a formula expression as for regression models, of the form
\code{response ~ predictors}. The response should be a factor or a
matrix with K columns, which will be interpreted as counts for each of
K classes.
A log-linear model is fitted, with coefficients zero for the first
class. An offset can be included: it should be a numeric matrix with K columns
if the response is either a matrix with K columns or a factor with K >= 2
classes, or a numeric vector for a response factor with 2 levels.
See the documentation of \code{\link{formula}()} for other details.
}

\item{data}{
an optional data frame in which to interpret the variables occurring
in \code{formula}.
}

\item{weights}{
optional case weights in fitting.
}

\item{subset}{
expression saying which subset of the rows of the data should  be used
in the fit. All observations are included by default.
}

\item{na.action}{
a function to filter missing data.
}

\item{contrasts}{
a list of contrasts to be used for some or all of
the factors appearing as variables in the model formula.
}

\item{model}{
  logical. If true, the model frame is saved as component \code{model}
  of the returned object.
}

\item{control}{a list of parameters for controlling the fitting
process. See \code{\link{brglmControl}} for details.}

\item{...}{arguments to be used to form the default 'control'
argument if it is not supplied directly.}
}
\description{
\code{bracl} is a wrapper of \code{\link{brglmFit}} that fits
adjacent category logit models using implicit and explicit bias
reduction methods. See Kosmidis & Firth (2011) for details.
}
\details{
Complete me
}
\examples{


}
\references{
Agresti A. (2002). Categorical data analysis (2nd
Edition). Wiley. New York.

Albert A. and Anderson J. A. (1984). On the Existence of Maximum
Likelihood Estimates in Logistic Regression Models. *Biometrika*,
**71** 1--10.

Kosmidis I. and Firth D. (2011). Multinomial logit bias reduction via
the Poisson log-linear model. *Biometrika*, **98**, 755-759.

Palmgren, J. (1981). The Fisher Information Matrix for Log Linear
Models Arguing Conditionally on Observed Explanatory
Variables. *Biometrika*, **68**, 563-566.
}
\seealso{
\code{\link[nnet]{multinom}}, \code{\link[nnet]{brmultinom}}
}
\author{
Ioannis Kosmidis \email{ioannis.kosmidis@warwick.ac.uk}
}
