% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brglm0.R
\name{brglm0}
\alias{brglm0}
\alias{brglm0.fit}
\alias{print.brglm0}
\alias{print.summary.brglm0}
\alias{summary.brglm0}
\title{Bias reduction in Binomial-response GLMs}
\usage{
brglm0(formula, family = binomial, data, weights, subset, na.action,
  start = NULL, etastart, mustart, offset, control.glm = glm.control1(...),
  model = TRUE, method = "brglm.fit", pl = FALSE, x = FALSE, y = TRUE,
  contrasts = NULL, control.brglm = brglm0.control(...), ...)
}
\description{
Fits binomial-response GLMs using the bias-reduction method developed in
Firth (1993) for the removal of the leading (\eqn{\mathop{\rm
O}(n^{-1})}{O(n^{-1})}) term from the asymptotic expansion of the bias
of the maximum likelihood estimator. Fitting is performed using
pseudo-data representations, as described in Kosmidis (2007, Chapter 5). For
estimation in binomial-response GLMs, the bias-reduction method is an
improvement over traditional maximum likelihood because:
\itemize{
\item the bias-reduced estimator is second-order unbiased and has smaller
      variance than the maximum likelihood estimator and
\item the resultant estimates and their corresponding standard errors
      are \bold{always} finite while the maximum likelihood estimates
      can be infinite (in situations where complete or quasi separation
      occurs).
}
}

