% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brglm2-defunct.R
\name{brglm2-defunct}
\alias{brglm2-defunct}
\alias{check_infinite_estimates}
\alias{detect_separation}
\title{Defunct Functions in package \pkg{brglm2}}
\usage{
check_infinite_estimates(object, ...)

detect_separation(
  x,
  y,
  weights = rep(1, nobs),
  start = NULL,
  etastart = NULL,
  mustart = NULL,
  offset = rep(0, nobs),
  family = gaussian(),
  control = list(),
  intercept = TRUE,
  singular.ok = TRUE
)
}
\description{
The functions or variables listed here are no longer part of
 \pkg{brglm2}.
}
\details{
\itemize{

\item \code{\link{detect_separation}}: This function is defunct
from \pkg{brglm2} since version 0.8.0. A new version of
\code{detect_separation} is now maintained in the
\pkg{detectseparation} R package at
\url{https://cran.r-project.org/package=detectseparation}.

\item \code{check_infinite_estimates} is defunct from
\pkg{brglm2} since version 0.8.0. An new version of
\code{check_infinite_estimates} is now maintained in the
\pkg{detectseparation} R package at
\url{https://cran.r-project.org/package=detectseparation}.

}
}
