% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brglm-package.r
\docType{package}
\name{brglm2}
\alias{brglm2}
\alias{brglm2-package}
\title{brglm2: Estimation and inference from generalized linear models
   using explicit and implicit methods for bias reduction}
\description{
Estimation and inference from generalized linear models based on
implicit methods for bias reduction (see Kosmidis, 2014, WIRE
computational statistics). brglm can achieve reduction of
estimation bias either through the adjusted score equations
approach in Firth (1993, Biometrika) and Kosmidis and Firth (2009,
Biometrika), or through the direct subtraction of an estimate of
the bias of the maximum likelihood estimator from the maximum
likelihood estimates. In the special case of generalized linear
models for binomial and multinomial responses, the adjusted score
equations approach returns estimates with improved frequentist
properties, that are also always finite, even in cases where the
maximum likelihood estimates are infinite (e.g. complete and
quasi-complete separation). Estimation in all cases takes place via
a modified Fisher scoring algorithm, and S3 methods for the
construction of of confidence intervals for the reduced-bias
estimates are provided.
}

