% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brglmControl.R
\name{brglmControl}
\alias{brglmControl}
\title{Auxiliary function for \code{\link{glm}} fitting using the
\code{\link{brglmFit}} method.  Typically only used internally by
\code{\link{brglmFit}}, but may be used to construct a
\code{control} argument to either function.}
\usage{
brglmControl(epsilon = 1e-10, maxit = 100, trace = FALSE,
  type = c("adjusted_scores", "correction", "maximum_likelihood"),
  dispTrans = "identity", slowit = 1, maxStepFactor = 1)
}
\arguments{
\item{epsilon}{positive convergence tolerance epsilon}

\item{maxit}{integer giving the maximal number of iterations
allowed}

\item{trace}{logical indicating if output should be produced for
each iteration}

\item{type}{type of estimation method. Supported methods include
\code{adjusted_scores}, \code{correction} and
\code{maximum_likelihood}}

\item{dispTrans}{the transformation of the dispersion to be
estimated. Default is \code{identity}. See Details.}

\item{slowit}{a positive real used as a multiplier for the
stepsize. The smaller it is the smaller the steps are}

\item{maxStepFactor}{the maximum number of step halving steps to
consider}
}
\value{
a list with components named as the arguments, including
    symbolic expressions for the dispersion transformation
    (\code{Trans}) and its inverse (\code{inverseTrans})
}
\description{
Auxiliary function for \code{\link{glm}} fitting using the
\code{\link{brglmFit}} method.  Typically only used internally by
\code{\link{brglmFit}}, but may be used to construct a
\code{control} argument to either function.
}
\details{
\code{\link{brglmControl}} provides default values and
    sanity checking for the various constants that control the
    iteration and generally the behaviour of
    \code{\link{brglmFit}}.

     When \code{trace} is true, calls to \code{cat} produce the
     output for each iteration.  Hence, \code{options(digits = *)}
     can be used to increase the precision.
}
\seealso{
\code{\link{brglmFit}} and \code{\link{glm.fit}}
}

