% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brglm2-package.R
\name{check_infinite_estimates}
\alias{check_infinite_estimates}
\title{Generic method for checking for infinite estimates}
\usage{
check_infinite_estimates(object, ...)
}
\arguments{
\item{object}{a fitted model object (e.g. the result of a
\code{\link{glm}} call).}

\item{...}{other options to be passed to the method.}
}
\description{
Generic method for checking for infinite estimates
}
\note{
\code{check_infinite_estimates} will be removed from \pkg{brglm2}
at version 0.8. An new version of
\code{check_infinite_estimates} is now maintained in the
\pkg{detectseparation} R package at
\url{https://cran.r-project.org/package=detectseparation}.
}
\seealso{
check_infinite_estimates.glm
}
