% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brglmFit.R
\name{coef.brglmFit}
\alias{coef.brglmFit}
\title{Extract model coefficients from \code{\link[=brglmFit]{"brglmFit"}} objects}
\usage{
\method{coef}{brglmFit}(object, model = c("mean", "full", "dispersion"), ...)
}
\arguments{
\item{object}{an object for which the extraction of model coefficients is
    meaningful.}

\item{model}{one of \code{"mean"} (default), \code{"dispersion"}, `"full",
to return the estimates of the parameters in the linear
prediction only, the estimate of the dispersion parameter only,
or both, respectively.}

\item{...}{other arguments.}
}
\description{
Extract model coefficients from \code{\link[=brglmFit]{"brglmFit"}} objects
}
\details{
See \code{\link[=coef]{coef()}} for more details.
}
\seealso{
\code{\link[=coef]{coef()}}
}
