% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expo.R
\name{coef.brglmFit_expo}
\alias{coef.brglmFit_expo}
\title{Extract estimates from \code{\link[=brglmFit_expo]{"brglmFit_expo"}} objects}
\usage{
\method{coef}{brglmFit_expo}(object, ...)
}
\arguments{
\item{object}{an object for which the extraction of model coefficients is
    meaningful.}

\item{...}{other arguments.}
}
\description{
Extract estimates from \code{\link[=brglmFit_expo]{"brglmFit_expo"}} objects
}
