% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brnb.R
\name{coef.brnb}
\alias{coef.brnb}
\title{Extract model coefficients from \code{\link{brnb}} objects}
\usage{
\method{coef}{brnb}(object, model = c("mean", "full", "dispersion"), ...)
}
\arguments{
\item{object}{an object for which the extraction of model coefficients is
    meaningful.}

\item{model}{one of \code{"mean"} (default), \code{"full"}, \code{"dispersion"},
to return the estimates of the parameters in the linear
prediction only, or both, the estimate of the dispersion parameter only,
 respectively.}

\item{...}{other arguments.}
}
\description{
Extract model coefficients from \code{\link{brnb}} objects
}
\details{
See \code{\link{coef}} for more details.
}
