% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmultinom.R
\name{confint.brmultinom}
\alias{confint.brmultinom}
\title{Method for computing confidence intervals for one or more
regression parameters in a \code{\link{brmultinom}} object}
\usage{
\method{confint}{brmultinom}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{a fitted model object.}

\item{parm}{a specification of which parameters are to be given
    confidence intervals, either a vector of numbers or a vector of
    names.  If missing, all parameters are considered.}

\item{level}{the confidence level required.}

\item{...}{additional argument(s) for methods.}
}
\description{
Method for computing confidence intervals for one or more
regression parameters in a \code{\link{brmultinom}} object
}
