% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brnb.R
\name{confint.brnb}
\alias{confint.brnb}
\title{Method for computing Wald confidence intervals for one or more
regression parameters in a \code{\link[=brnb]{"brnb"}} object}
\usage{
\method{confint}{brnb}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{a fitted model object.}

\item{parm}{a specification of which parameters are to be given
    confidence intervals, either a vector of numbers or a vector of
    names.  If missing, all parameters are considered.}

\item{level}{the confidence level required.}

\item{...}{additional argument(s) for methods.}
}
\description{
Method for computing Wald confidence intervals for one or more
regression parameters in a \code{\link[=brnb]{"brnb"}} object
}
