% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brglm2-package.R
\name{detect_infinite_estimates}
\alias{detect_infinite_estimates}
\title{Generic method for detecting infinite estimates}
\usage{
detect_infinite_estimates(object, ...)
}
\arguments{
\item{object}{a fitted model object (e.g. the result of a
\code{\link{glm}} call)}

\item{...}{other options to be passed to the method}
}
\description{
Generic method for detecting infinite estimates
}

