% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_separation.R, R/zzz_conventions.R
\name{detect_separation_control}
\alias{detect_separation_control}
\alias{detectSeparationControl}
\title{Auxiliary function for the \code{\link{glm}} interface when
\code{method} is \code{\link{detect_separation}}.}
\usage{
detect_separation_control(linear_program = c("primal", "dual"),
  purpose = c("find", "test"),
  beta_tolerance = sqrt(.Machine$double.eps))

detectSeparationControl(linear_program = c("primal", "dual"),
  purpose = c("find", "test"),
  beta_tolerance = sqrt(.Machine$double.eps))
}
\arguments{
\item{linear_program}{should \code{\link{detect_separation}} solve
the \code{"primal"} or \code{"dual"} linear program for
separation detection?}

\item{purpose}{should \code{\link{detect_separation}} simply
\code{"test"} for separation or also \code{"find"} which
parameters are infinite?}

\item{beta_tolerance}{maximum absolute variable value from the
linear program, before separation is declared.}
}
\description{
Typically only used internally by \code{\link{detect_separation}}
but may be used to construct a \code{control} argument.
}
