% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brglm-package.r
\name{enrich}
\alias{enrich}
\title{Generic method for enriching objects}
\usage{
enrich(object, ...)
}
\arguments{
\item{object}{the object to be enriched}

\item{...}{Arguments to be passed to other methods}
}
\description{
Generic method for enriching objects
}

