% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdyplFit.R
\name{hd_summary.mdyplFit}
\alias{hd_summary.mdyplFit}
\title{High-dimensionality correction of estimates and Wald statistics
from \code{\link[=mdyplFit]{mdyplFit()}} objects.}
\usage{
hd_summary.mdyplFit(object, se_start, null = 0, ...)
}
\arguments{
\item{object}{an \code{\link[=mdyplFit]{"mdyplFit"}} object.}

\item{se_start}{starting values for the parameters of the state
evoultion equations; see \code{\link[=solve_se]{solve_se()}}.}

\item{null}{a vector of null values for the parameters estimated in
\code{object}. Default is \code{0}.}
}
\value{
A table of corrected coefficients, Wald statistics and p-values
}
\description{
High-dimensionality correction of estimates and Wald statistics
from \code{\link[=mdyplFit]{mdyplFit()}} objects.
}
