% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmultinom.R
\name{print.summary.brmultinom}
\alias{print.summary.brmultinom}
\title{Print method for \code{\link{summary.brmultinom}} objects}
\usage{
\method{print}{summary.brmultinom}(x, digits = x$digits, ...)
}
\description{
Print method for \code{\link{summary.brmultinom}} objects
}
\section{Note}{

Code adopted from \code{nnet:::print.summary.brmultinom}
}

