% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmultinom.R
\name{simulate.brmultinom}
\alias{simulate.brmultinom}
\title{Method for simulating data sets from \code{\link{brmultinom}} and
\code{\link{bracl}} objects}
\usage{
\method{simulate}{brmultinom}(object, ...)
}
\description{
Method for simulating data sets from \code{\link{brmultinom}} and
\code{\link{bracl}} objects
}
