% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stemcell}
\alias{stemcell}
\title{Opinion on Stem Cell Research and Religious Fundamentalism}
\format{An object of class \code{data.frame} with 24 rows and 4 columns.}
\source{
The \code{stemcell} data set is analysed in Agresti (2010, Subsection~4.1.5).
}
\usage{
stemcell
}
\description{
A data set from the 2006 General Social Survey that shows the
relationship in the United States between opinion about funding
stem cell research and the fundamentalism/liberalism of one’s
religious beliefs, stratified by gender.
}
\references{
Agresti, A. (2010). *Analysis of Ordinal Categorical Data* (2nd edition).  Wiley Series in
Probability and Statistics. Wiley
}
\keyword{datasets}
