% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unless_null}
\alias{unless_null}
\title{A simple utility that checks whether its argument x is null and
returns if.null if it is and x if it is not}
\usage{
unless_null(x, if_null)
}
\description{
A simple utility that checks whether its argument x is null and
returns if.null if it is and x if it is not
}

