% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brglm2-defunct.R
\name{brglm2-defunct}
\alias{brglm2-defunct}
\alias{check_infinite_estimates}
\alias{detect_separation}
\title{Defunct Functions in package \pkg{brglm2}}
\usage{
check_infinite_estimates(...)

detect_separation(...)
}
\arguments{
\item{...}{arguments to be passed to functions and methods.}
}
\description{
The functions or variables listed here are no longer part of
 \pkg{brglm2}.
}
\details{
\itemize{

\item \code{\link{detect_separation}}: This function is defunct
from \pkg{brglm2} since version 0.8.0. A new version of
\code{detect_separation} is now maintained in the
\pkg{detectseparation} R package at
\url{https://cran.r-project.org/package=detectseparation}.

\item \code{check_infinite_estimates} is defunct from
\pkg{brglm2} since version 0.8.0. An new version of
\code{check_infinite_estimates} is now maintained in the
\pkg{detectseparation} R package at
\url{https://cran.r-project.org/package=detectseparation}.

}
}
