% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brglmFit.R
\name{brglmFit}
\alias{brglmFit}
\title{Fitting function for \code{\link{glm}} that allows reduced-bias estimation and inference}
\usage{
brglmFit(x, y, weights = rep(1, nobs), start = NULL, etastart = NULL,
  mustart = NULL, offset = rep(0, nobs), family = gaussian(),
  control = list(), intercept = TRUE)
}
\arguments{
\item{x}{For \code{glm}:
    logical values indicating whether the response vector and model
    matrix used in the fitting process should be returned as components
    of the returned value.

    For \code{glm.fit}: \code{x} is a design matrix of dimension
    \code{n * p}, and \code{y} is a vector of observations of length
    \code{n}.
  }

\item{weights}{an optional vector of \sQuote{prior weights} to be used
    in the fitting process.  Should be \code{NULL} or a numeric vector.}

\item{start}{starting values for the parameters in the linear predictor.}

\item{etastart}{starting values for the linear predictor.}

\item{mustart}{starting values for the vector of means.}

\item{offset}{this can be used to specify an \emph{a priori} known
    component to be included in the linear predictor during fitting.
    This should be \code{NULL} or a numeric vector of length equal to
    the number of cases.  One or more \code{\link{offset}} terms can be
    included in the formula instead or as well, and if more than one is
    specified their sum is used.  See \code{\link{model.offset}}.}

\item{family}{a description of the error distribution and link
    function to be used in the model.  For \code{glm} this can be a
    character string naming a family function, a family function or the
    result of a call to a family function.  For \code{glm.fit} only the
    third option is supported.  (See \code{\link{family}} for details of
    family functions.)}

\item{control}{a list of parameters controlling the fitting
process. See \code{\link{brglmControl}} for details.}

\item{intercept}{logical. Should an intercept be included in the
    \emph{null} model?}
}
\description{
\code{\link{brglmFit}} is a fitting function for \code{\link{glm}}
(see the description of \code{method} and the \code{Fitting
functions} section in \code{\link{glm}} that fits generalized
linear models using implicit and explicit bias reduction
methods. Estimation is performed using a quasi-Fisher scoring
iteration based on the iterative correction of the bias of the
Fisher scoring iterates.
}
\details{
\itemize{ \item }
}
\examples{
\dontrun{
## The lizards example from ?brglm::brglm
load("../data/lizards.rda")
# Fit the model using maximum likelihood
lizardsML <- glm(cbind(grahami, opalinus) ~ height + diameter +
                     light + time, family = binomial(logit), data=lizards,
                 method = "glm.fit")
# Now the bias-reduced fit:
lizardsBR <- glm(cbind(grahami, opalinus) ~ height + diameter +
                     light + time, family = binomial(logit), data=lizards,
                 method = "brglmFit")
}


## An example from
## King, Gary, James E. Alt, Nancy Elizabeth Burns and Michael Laver
## (1990).  "A Unified Model of Cabinet Dissolution in Parliamentary
## Democracies", _American Journal of Political Science_, vol. 34,
## no. 3, pp. 846-870.

\dontrun{
library(Zelig)
data(coalition)
# The maximum likelihood fit with log link
fm1 <- glm(duration ~ fract + numst2, family = Gamma, data = coalition)
# The bias-reduced fit
fm1br <- update(fm1, method = "brglmFit")
# The bias-corrected fit
fm1bc <- update(fm1, method = "brglmFit", correction = TRUE)
}
}
\seealso{
\code{\link{glm.fit}} and \code{\link{glm}}
}

