% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brglmFit.R
\name{brglmFit}
\alias{brglmFit}
\title{Fitting function for \code{\link{glm}} for reduced-bias
estimation and inference}
\usage{
brglmFit(x, y, weights = rep(1, nobs), start = NULL, etastart = NULL,
  mustart = NULL, offset = rep(0, nobs), family = gaussian(),
  control = list(), intercept = TRUE, fixedTotals = NULL)
}
\arguments{
\item{x}{\code{x} is a design matrix of dimension \code{n * p},}

\item{y}{\code{y} is a vector of observations of length \code{n}.}

\item{weights}{an optional vector of \sQuote{prior weights} to be used
    in the fitting process.  Should be \code{NULL} or a numeric vector.}

\item{start}{starting values for the parameters in the linear
predictor. If \code{NULL} (default) then the maximum likelihood
estimates are caluclated and used as starting values}

\item{etastart}{applied only when start is not
\code{NULL}. Starting values for the linear predictor to be
passed to \code{\link{glm.fit}} when computing starting values
using maximum likelihood.}

\item{mustart}{applied only when start is not \code{NULL}. Starting
values for the vector of means to be passed to
\code{\link{glm.fit}} when computing starting values using
maximum likelihood.}

\item{offset}{this can be used to specify an \emph{a priori} known
    component to be included in the linear predictor during fitting.
    This should be \code{NULL} or a numeric vector of length equal to
    the number of cases.  One or more \code{\link{offset}} terms can be
    included in the formula instead or as well, and if more than one is
    specified their sum is used.  See \code{\link{model.offset}}.}

\item{family}{a description of the error distribution and link
    function to be used in the model.  For \code{glm} this can be a
    character string naming a family function, a family function or the
    result of a call to a family function.  For \code{glm.fit} only the
    third option is supported.  (See \code{\link{family}} for details of
    family functions.)}

\item{control}{a list of parameters controlling the fitting
process. See \code{\link{brglmControl}} for details.}

\item{intercept}{logical. Should an intercept be included in the
    \emph{null} model?}

\item{fixedTotals}{Effective only when \code{family} is
\code{poisson}. Either \code{NULL} (no effect) or a vector that
indicates which counts must be treated as a group. See
Details for more information and \code{\link{brmultinom}}.}

\item{...}{arguments to be used to form the default 'control'
argument if it is not supplied directly.}
}
\description{
\code{\link{brglmFit}} is a fitting function for \code{\link{glm}}
that fits generalized linear models using implicit and explicit
bias reduction methods. Currently supported methods include the
implicit adjusted scores approach in Firth (1993) and Kosmidis \&
Firth (2009), the correction of the asymptotic bias in Cordeiro &
McCullagh (1991), and maximum likelihood.  Estimation is performed
using a quasi-Fisher scoring iteration based on the iterative
correction of the asymptotic bias of the Fisher scoring iterates.
}
\details{
Below
\itemize{
\item The null deviance is evaluated based on estimates of the
    expectations using the bias reduction method specified by the
    \code{type} argument (see \code{\link{brglmControl}}).

\item See the description of \code{method} and the \code{Fitting
functions} section in \code{\link{glm}}
#'
\item If \code{type == "correction"} (see
\code{\link{brglmControl}}), then \code{coefficients} and
\code{transDispersion} carry the estimated biases as attributes.
}
}
\examples{
## The lizards example from ?brglm::brglm
data("lizards")
# Fit the model using maximum likelihood
lizardsML <- glm(cbind(grahami, opalinus) ~ height + diameter +
                     light + time, family = binomial(logit), data = lizards,
                 method = "glm.fit")
# Now the bias-reduced fit:
lizardsBR <- glm(cbind(grahami, opalinus) ~ height + diameter +
                     light + time, family = binomial(logit), data = lizards,
                 method = "brglmFit")
summary(lizardsML)
summary(lizardsBR)


## Another example from
## King, Gary, James E. Alt, Nancy Elizabeth Burns and Michael Laver
## (1990).  "A Unified Model of Cabinet Dissolution in Parliamentary
## Democracies", _American Journal of Political Science_, vol. 34,
## no. 3, pp. 846-870.

\dontrun{
data("coalition", package = "Zelig")
# The maximum likelihood fit with log link
coalitionML <- glm(duration ~ fract + numst2, family = Gamma, data = coalition)
# The bias-reduced fit
coalitionBR <- update(coalitionML, method = "brglmFit")
# The bias-corrected fit
coalitionBC <- update(coalitionML, method = "brglmFit", type = "correction")
}

\dontrun{
## An example with offsets from Venables & Ripley (2002, p.189)
data("anorexia", package = "MASS")

anorexML <- glm(Postwt ~ Prewt + Treat + offset(Prewt),
                family = gaussian, data = anorexia)
anorexBR <- update(anorexML, method = "brglmFit")
anorexBC <- update(anorexML, method = "brglmFit", type = "correction")

## The outputs are identical, because the maximum likelihood
## estimators of the regression parameters are unbiased when family
## is Gaussian, and the bias-reduced estimator of the dispersion is
## the unbiased, by degree of freedom adjustment, estimator of the
## residual variance.
summary(anorexML)
summary(anorexBR)
summary(anorexBC, start = coef(anorexML))
}

}
\references{
A
}
\seealso{
\code{\link{glm.fit}} and \code{\link{glm}}
}

