% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmultinom.R
\name{brmultinom}
\alias{brmultinom}
\title{Bias reduction for multinomial response models using the
"Poisson trick".}
\usage{
brmultinom(formula, data, weights, subset, na.action, contrasts = NULL,
  control = list(...), ...)
}
\arguments{
\item{formula}{
a formula expression as for regression models, of the form
\code{response ~ predictors}. The response should be a factor or a
matrix with K columns, which will be interpreted as counts for each of
K classes.
A log-linear model is fitted, with coefficients zero for the first
class. An offset can be included: it should be a numeric matrix with K columns
if the response is either a matrix with K columns or a factor with K >= 2
classes, or a numeric vector for a response factor with 2 levels.
See the documentation of \code{\link{formula}()} for other details.
}

\item{data}{
an optional data frame in which to interpret the variables occurring
in \code{formula}.
}

\item{weights}{
optional case weights in fitting.
}

\item{subset}{
expression saying which subset of the rows of the data should  be used
in the fit. All observations are included by default.
}

\item{na.action}{
a function to filter missing data.
}

\item{contrasts}{
a list of contrasts to be used for some or all of
the factors appearing as variables in the model formula.
}

\item{control}{a list of parameters for controlling the fitting
process. See \code{\link{brglmControl}} for details.}

\item{...}{arguments to be used to form the default 'control'
argument if it is not supplied directly.}
}
\description{
\code{brmultinom} is a wrapper of \code{\link{brglmFit}} that fits
multinomial regression models using implicit and explicit bias
reduction methods. See Kosmidis & Firth (2011) for details.
}
\details{
\code{brmultinom} is a wrapper of \code{\link{brglmFit}} that fits
multinomial regression models through the "Poisson trick" (see, for
example, Palmgren, 1981 and Kosmidis & Firth, 2009). The
implementation relies on the construction of an "extended" model
matrix for the log-linear model and constraints on the sums of the
Poisson means. Specifically, a log-linear model is fitted on a
Kronecker product
(\url{https://en.wikipedia.org/wiki/Kronecker_product}) of the
original model matrix \code{X} implied by the formula, augmented by
\code{nrow(X)} dummy variables.

The extended model matrix is sparse, and the \pkg{Matrix} package
is used for its effective storage.

While \code{\link{brmultinom}} can be used for analyses using
multinomial regression models, the current implementation is more
of a "proof of concept" and is not expected to scale well with
either of \code{nrow(X)}, \code{ncol(X)} or the number of levels in
the cateogrical response.
}
\seealso{
\code{\link[nnet]{multinom}}
}

