% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coalition}
\alias{coalition}
\title{Coalition data}
\format{
An object of class \code{data.frame} with 314 rows and 7 columns.
}
\usage{
coalition
}
\description{
Coalition data
}
\note{
Data is as provided in the Zeilig R package (\url{https://cran.r-project.org/package=Zelig})
}
\references{
King G, Alt J E, Burns N E, Laver M. (1990). A Unified
 Model of Cabinet Dissolution in Parliamentary
 Democracies. *American Journal of Political Science*, **34**,
 846-870. \url{https://doi.org/10.2307/2111401}

 King G, Alt J E, Burns N E, Laver M. ICPSR
 Publication Related Archive, 1115.
}
\seealso{
\code{\link{brglm_fit}}
}
\keyword{datasets}
