% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{enrich.family}
\alias{enrich.family}
\title{Enrich family object with family- and link-specfic mathematical functions useful to higher-order
asymptotic methods for generalized linear models}
\usage{
\method{enrich}{family}(object)
}
\arguments{
\item{object}{an object of class \link{class{family}}.}
}
\description{
Enrich family object with family- and link-specfic mathematical functions useful to higher-order
asymptotic methods for generalized linear models
}
\details{
Enriches the family object with the second and third derivatives of
the inverse link function with respect to \code{eta}
(\code{d2mu.eta} and \code{d3mu.eta}; the first is \code{mu.eta}),
the first and second derivative of the variance function with
respect to \code{mu} (\code{d1variance}, \code{d2variance}), and
the first three derivatives of


}

