% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdyplFit.R
\name{sloe}
\alias{sloe}
\title{Estimate the corrupted signal strength in a model with
(sub-)Gaussian covariates}
\usage{
sloe(object)
}
\arguments{
\item{object}{an \code{\link[=mdyplFit]{"mdyplFit"}} object.}
}
\description{
Estimate the corrupted signal strength in a model with
(sub-)Gaussian covariates
}
\details{
The Signal Strength Leave-One-Out Estimator (SLOE) is defined in
Yadlowsky et al. (2021) when the model is estimated using maximum
likelihood. The SLOE adaptation when estimation is through maximum
Diaconis-Ylvisaker prior penalized likelihood (\code{\link[=mdypl_fit]{mdypl_fit()}}) has
been done in Sterzinger & Kosmidis (2025).
}
\references{
Yadlowsky S, Yun T, McLean C, D'Amour A (2021). SLOE: a faster
method for statistical inference in high-dimensional logistic
regression. In
"Proceedings of the 35th International Conference on Neural Information Processing Systems",
NIPS '21. Curran Associates Inc., Red Hook, NY, USA. ISBN
9781713845393.
}
