% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brnb.R
\name{summary.brnb}
\alias{summary.brnb}
\alias{print.summary.brnb}
\title{\code{\link[=summary]{summary()}} method for \code{\link[=brnb]{"brnb"}} objects}
\usage{
\method{summary}{brnb}(object, ...)

\method{print}{summary.brnb}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{an object of class \code{\link[=brnb]{"brnb"}}, typically, a
result of a call to \code{\link[=brnb]{brnb()}}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object of class \code{\link[=summary.brnb]{"summary.brnb"}},
usually, a result of a call to \link{summary.brnb}.}

\item{digits}{the number of significant digits to use when printing.}
}
\description{
\code{\link[=summary]{summary()}} method for \code{\link[=brnb]{"brnb"}} objects
}
\details{
The interface of the summary method for \code{\link[=brnb]{"brnb"}}
objects is similar to that of \code{\link[=brglmFit]{"brglmFit"}} objects
with additional information.

\if{html}{\out{<div class="sourceCode">}}\preformatted{p-values of the individual Wald statistics are based on the
standard normal distribution.
}\if{html}{\out{</div>}}
}
\examples{
# For examples see examples(brnb)

}
\seealso{
\code{\link[=summary.brglmFit]{summary.brglmFit()}} and \code{\link[=glm]{glm()}}
}
