% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brglmFit.R
\name{vcov.brglmFit}
\alias{vcov.brglmFit}
\title{Return the variance-covariance matrix for the regression parameters
in a \code{\link{brglmFit}} object}
\usage{
\method{vcov}{brglmFit}(object, model = c("mean", "full", "dispersion"), ...)
}
\arguments{
\item{object}{
    a fitted model object, typically.  Sometimes also a
    \code{\link{summary}()} object of such a fitted model.
  }

\item{model}{character specifying for which component of the model coefficients should be extracted.}

\item{...}{
    additional arguments for method functions.  For the
    \code{\link[stats]{glm}} method this can be used to pass a
    \code{dispersion} parameter.}
}
\description{
Return the variance-covariance matrix for the regression parameters
in a \code{\link{brglmFit}} object
}
